Compilation warnings removed + a few fixes.

+++ byteorder.c
@@ -48,7 +48,7 @@ int main(int argc, char **argv)
 			"-h	help\n");
 		return 0;
 	}
-		
+
 	x = (unsigned char*) &test;
 
 	if (*x == 0x00) {
@@ -68,7 +68,8 @@ int main(int argc, char **argv)
 	}
 
 	if (debug) {
-		printf("sizeof(unsigned int) = %d\n", sizeof(unsigned int));
+		printf("sizeof(unsigned int) = %d\n",
+                    (int)sizeof(unsigned int));
 		printf("unsigned int test = 1;\n");
 		printf("in memory as: ");
 		for (j = 0; j < sizeof(unsigned int); j++)
